use molecule for serialize and deserialize hopdata and tlcerr
