+++ Cargo.toml
@@ -63,6 +63,7 @@ tokio = { version = "1", features = [
 indicatif = "0.16"
 console = "0.15.8"
 bincode = "1.3.3"
+num_enum = "0.7.3"
 
 [features]
 default = []
