+++ src/fiber/schema/fiber.mol
@@ -1,5 +1,6 @@
 import blockchain;
 
+array Uint16 [byte; 2];
 array EcdsaSignature [byte; 64];
 array PubNonce [byte; 66]; // PubNonce used by musig2 crate.
 option PubNonceOpt (PubNonce);
@@ -186,4 +187,53 @@ union FiberMessage {
     ClosingSigned,
     ReestablishChannel,
     AnnouncementSignatures,
+}
+
+option PaymentPreimageOpt (Byte32);
+option PubkeyOpt (Pubkey);
+table PaymentHopData {
+    amount: Uint128,
+    expiry: Uint64,
+    payment_preimage: PaymentPreimageOpt,
+    hash_algorithm: byte,
+    funding_tx_hash: Byte32,
+    next_hop: PubkeyOpt,
+}
+
+// A ChannelUpdate is a update to the public channel. Newer channel information may be updated with
+// a ChannelUpdate with a larger timestamp.
+struct ChannelUpdate {
+    // Signature of the node that wants to update the channel information.
+    signature: EcdsaSignature,
+    chain_hash: Byte32,
+    channel_outpoint: OutPoint,
+    timestamp: Uint64,
+    message_flags: Uint32,
+    channel_flags: Uint32,
+    tlc_expiry_delta: Uint64,
+    tlc_minimum_value: Uint128,
+    tlc_maximum_value: Uint128,
+    tlc_fee_proportional_millionths: Uint128,
+}
+
+option ChannelUpdateOpt (ChannelUpdate);
+table ChannelFailed {
+    channel_outpoint: OutPoint,
+    channel_update: ChannelUpdateOpt,
+    node_id: Pubkey,
+}
+
+table NodeFailed {
+    node_id: Pubkey,
+}
+
+union TlcErrData {
+    ChannelFailed,
+    NodeFailed,
+}
+
+option TlcErrDataOpt (TlcErrData);
+table TlcErr {
+    error_code: Uint16,
+    extra_data: TlcErrDataOpt,
 }
